requirements.txt
keyboard==0.13.4
pillow==6.2.1

